requirements.txt
streamlit==1.29.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
python-telegram-bot==20.7
transformers==4.36.0
torch
datasets==2.15.0
scikit-learn==1.3.2